Add guard to models. Add policy to user functions. Add super admin ability

+++ app/Http/Controllers/Api/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
 use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -14,11 +15,14 @@ class UsersController extends Controller
     /**
      * Display a listing of the resource.
      *
+     * @param Request $request
      * @return Response
+     * @throws AuthorizationException
      */
     public function index(Request $request)
     {
-        $users = User::latest()->paginate(25);
+        $this->authorize('viewAny');
+        $users = User::where('access_level', '<', Auth::user()->access_level)->latest()->paginate(25);
 
         return $users;
     }
@@ -29,10 +33,16 @@ class UsersController extends Controller
      * @param Request $request
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function store(Request $request)
     {
-        
+        $this->authorize('create');
+
+        if ($request->access_level && $request->access_level >= Auth::user()->access_level) {
+            return response()->json(['error' => 'Unauthorized - You cannot create users of the same access level or higher'], 401);
+        }
+
         $user = User::create($request->all());
 
         return response()->json($user, 201);
@@ -41,13 +51,15 @@ class UsersController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function show($id)
     {
         $user = User::findOrFail($id);
+        $this->authorize('view', $user);
 
         return $user;
     }
@@ -56,14 +68,15 @@ class UsersController extends Controller
      * Update the specified resource in storage.
      *
      * @param Request $request
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(Request $request, $id)
     {
-        
         $user = User::findOrFail($id);
+        $this->authorize('update', $user);
         $user->update($request->all());
 
         return response()->json($user, 200);
@@ -72,12 +85,15 @@ class UsersController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $user = User::findOrFail($id);
+        $this->authorize('delete', $user);
         User::destroy($id);
 
         return response()->json(null, 204);
